add hunting file and import into trail file
